requirements.txt
chameleon==2.24
decorator==4.0.6
ipython==4.1.0
ipython-genutils==0.1.0
jinja2==2.8
mako==1.0.3
markdown==2.6.5
markupsafe==0.23
passlib==1.6.5
PasteDeploy==1.5.2
path.py==8.1.2
pickleshare==0.6
pygments==2.1
pyramid==1.6
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4.2
pyramid-ipython==0.1
pyramid-jinja2==2.6.2
pyramid-mako==1.0.2
pyramid-tm==0.12.1
repoze.lru==0.6
simplegeneric==0.8.1
sqlalchemy==1.0.11
traitlets==4.1.0
transaction==1.4.4
translationstring==1.3
venusian==1.0
waitress==0.8.10
WebOb==1.5.1
wtforms==2.1
zope.deprecation==4.1.2
zope.interface==4.1.3
zope.sqlalchemy==0.7.6